modularize pwm
